stop clobbering my version.json!
- turns out that a static version.json was added to the docker folder at
some point; that file took precedence over the one from circle. Problem
is, the static version.json had an invalid build host and our
verification scripts refused to commit the image to the correct internal
repo. All of this is fixed now.
- clean up makefile
- remove static version.json
- copy the circle version.json to the correct place

use a consistent port in all the places